+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Minor Changes
 
+- 1078 | 2 files where without explicit error
 - 988 | [Digital service] Edit a value with a decimal reset the values to 0
 - 985 | Issue while exporting virtual equipment with new architecture